Refresh merged integer boundary goldens

+++ tests/integration/cases/int_list_signed_18_digit_boundaries/Go@v1_18.go
@@ -1,7 +1,7 @@
 package main
 
 func main() {
-my_data := []int{
+my_data := []int64{
 	999999999999999999,
 	-999999999999999999,
 }
