+++ tests/integration/cases/int_list_signed_18_digit_boundaries/Go_combined_declaration_style_var@v1_18.go
@@ -1,11 +1,11 @@
 package main
 
 func main() {
-var my_data = []int{
+var my_data = []int64{
 	999999999999999999,
 	-999999999999999999,
 }
-my_data = []int{
+my_data = []int64{
 	999999999999999999,
 	-999999999999999999,
 }
